feat: update the script to not do anything if certs exist

+++ hack/create_developer_certificate_secrets.sh
@@ -22,14 +22,15 @@ certPath="./hack/certs/cert.pem"
 keyPath="./hack/certs/key.pem"
 rootCAPath="./hack/certs/rootCA.pem"
 
-echo "updating root certificate"
-
-sudo cat "${rootCAPath}" | sudo tee -a /etc/ssl/certs/ca-certificates.crt || echo "failed to append to ca-certificates. Ignoring the failure"
-
 if [ ! -e "${certPath}" ] && [ ! -e "${keyPath}" ]; then
   echo -n "certificates not found, generating..."
 
   CAROOT=./hack/certs ./bin/mkcert -cert-file ./hack/certs/cert.pem -key-file ./hack/certs/key.pem registry.ocm-system.svc.cluster.local localhost 127.0.0.1 ::1
+
+  if [ -e '/etc/ssl/certs/ca-certificates.crt' ]; then
+    echo "updating root certificate"
+    sudo cat "${rootCAPath}" | sudo tee -a /etc/ssl/certs/ca-certificates.crt || echo "failed to append to ca-certificates. Ignoring the failure"
+  fi
 
   echo "done"
 else
